updated to use sql server, added precision to decimals to rid warnings, migration completed

+++ Lewiss/Lewiss.Pricing.Data/Model/Product.cs
@@ -11,7 +11,7 @@ public class Product
     public required int Width { get; set; }
     public required int Height { get; set; }
     public required int Reveal { get; set; }
-    public required decimal InstallHeight { get; set; }
+    public required int InstallHeight { get; set; }
     public required int RemoteNumber { get; set; }
     public required int RemoteChannel { get; set; }
     public required int WorksheetId { get; set; }
